+++ experimental/packages/api-logs/src/types/LogRecord.ts
@@ -14,12 +14,10 @@
  * limitations under the License.
  */
 
-import { AttributeValue, Context, TimeInput } from '@opentelemetry/api';
+import { AnyValue, AnyValueMap, Context, TimeInput } from '@opentelemetry/api';
 
-export type LogAttributeValue = AttributeValue | LogAttributes;
-export interface LogAttributes {
-  [attributeKey: string]: LogAttributeValue | undefined;
-}
+export type LogAttributeValue = AnyValue;
+export type LogAttributes = AnyValueMap;
 
 export enum SeverityNumber {
   UNSPECIFIED = 0,